发送消息

+++ service-wx/service-wx-api/src/main/java/com/soybean/mall/wx/mini/order/bean/request/WxCreateAfterSaleRequest.java
@@ -30,7 +30,7 @@ public class WxCreateAfterSaleRequest implements Serializable {
     /**
      * 退款金额，单位：分
      */
-    private BigDecimal refund;
+    private Integer refund;
     /**
      * 售后类型，1:退款,2:退款退货,3:换货
      */
